refactor: migrate Chakra UI utilities to new library path
Moved Chakra UI utilities from `chakra-ui/ui` to `lib/chakra-ui/generated`. Updated all associated imports across the codebase for consistency. This improves project organization and aligns shared resources with the new library structure.
