+++ apps/nextjs/src/hooks/use-toaster.tsx
@@ -1,4 +1,4 @@
-import { toaster } from '@/chakra-ui/ui/toaster';
+import { toaster } from '@/lib/chakra-ui/generated/toaster';
 
 export const useToaster = () => {
   return {
